document-assistant/frontend/requirements.txt
streamlit
langchain-groq
langchain-community
langchain-core
huggingface-hub
faiss-cpu
pydantic
python-dotenv